修正获取毫秒级时间戳的接口

+++ 02_App/Src/BSW/SystemService/TM/SS_Tm.c
@@ -185,7 +185,7 @@ uint32_t SSTM_GetSecTimestamp(void)
     return g_stTmCtrl.secSysTimestamp;
 }
 
-uint32_t SSTM_GetMsecTimestamp(void)
+uint64_t SSTM_GetMsecTimestamp(void)
 {
     return g_stTmCtrl.msecSysTimeStamp;
 }
