+++ 02_App/Src/BSW/SystemService/TM/SS_Tm.h
@@ -46,7 +46,7 @@ void SSTM_SynTimeByDateTime(CommonDateTime_Struct *pTime);
 void SSTM_MainFunction(void);
 void SSTM_InitMemory(void);
 uint32_t SSTM_GetSecTimestamp(void);
-uint32_t SSTM_GetMsecTimestamp(void);
+uint64_t SSTM_GetMsecTimestamp(void);
 uint8_t SSTM_GetSyncTimeFlag(void);
 void SSTM_GetDateTime(CommonDateTime_Struct *pTime);
 void SSTM_GetTimeStr(char *pTimeStr);
